Starting to re-split validation.py
It was optimistic to think I could pull all validation into one file,
given the strict dependencies on conditions and models.

I also need to migrate the unit tests to their correct files.
